Second commit

+++ backend/controllers/product.controller.js
@@ -55,13 +55,18 @@ export const updateProduct = async(req, res) => {
 
 export const deleteProduct = async(req, res) => {
     const {id} = req.params
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({
+            success: false, message: 'Object not found'
+        });
+    }
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({ 
             success:true, message: "product deleted"})
     } catch (error) {
         console.log("error in deleting product", error.message);
-        res.status(404).json({ 
-            success: false, message: 'product not found'})
+        res.status(500).json({ 
+            success: false, message: 'Server Error'})
     }
 };
